set up d3 dynamic graph

+++ models/search.js
@@ -1,8 +1,8 @@
 module.exports = function(sequelize, DataTypes) {
   var Search = sequelize.define("Search", {
     search: DataTypes.STRING,
-    price: DataTypes.INTEGER
-    // description: DataTypes.TEXT
+    price: DataTypes.DECIMAL(15, 2)
   });
+
   return Search;
 };
